+++ src/module_library/c3_canopy.cpp
@@ -21,6 +21,10 @@ string_vector c3_canopy::get_inputs()
         "electrons_per_carboxylation",  // electron / carboxylation
         "electrons_per_oxygenation",    // electron / oxygenation
         "gbw_canopy",                   // m / s
+        "gm_at_25",                     // mol / m^2 / s / Pa
+        "gm_Ha",                        // J / mol
+        "gm_Hd",                        // J / mol
+        "gm_S",                         // J / K / mol
         "growth_respiration_fraction",  // dimensionless
         "Gs_min",                       // mol / m^2 / s
         "Gstar_at_25",                  // micromol / mol
@@ -88,6 +92,9 @@ void c3_canopy::do_operation() const
 {
     // Combine temperature response parameters
     c3_temperature_response_parameters const tr_param{
+        gm_Ha,
+        gm_Hd,
+        gm_S,
         Gstar_Ea,
         Jmax_Ea,
         Kc_Ea,
@@ -120,6 +127,7 @@ void c3_canopy::do_operation() const
         electrons_per_carboxylation,
         electrons_per_oxygenation,
         gbw_canopy,
+        gm_at_25,
         growth_respiration_fraction,
         Gs_min,
         Gstar_at_25,