+++ src/module_library/c3_leaf_photosynthesis.h
@@ -33,6 +33,10 @@ class c3_leaf_photosynthesis : public direct_module
           electrons_per_carboxylation{get_input(input_quantities, "electrons_per_carboxylation")},
           electrons_per_oxygenation{get_input(input_quantities, "electrons_per_oxygenation")},
           gbw_canopy{get_input(input_quantities, "gbw_canopy")},
+          gm_at_25{get_input(input_quantities, "gm_at_25")},
+          gm_Ha{get_input(input_quantities, "gm_Ha")},
+          gm_Hd{get_input(input_quantities, "gm_Hd")},
+          gm_S{get_input(input_quantities, "gm_S")},
           Gs_min{get_input(input_quantities, "Gs_min")},
           Gstar_at_25{get_input(input_quantities, "Gstar_at_25")},
           Gstar_Ea{get_input(input_quantities, "Gstar_Ea")},
@@ -65,6 +69,7 @@ class c3_leaf_photosynthesis : public direct_module
 
           // Get pointers to output quantities
           Assim_op{get_op(output_quantities, "Assim")},
+          Cc_op{get_op(output_quantities, "Cc")},
           Ci_op{get_op(output_quantities, "Ci")},
           Cs_op{get_op(output_quantities, "Cs")},
           EPenman_op{get_op(output_quantities, "EPenman")},
@@ -100,6 +105,10 @@ class c3_leaf_photosynthesis : public direct_module
     double const& electrons_per_carboxylation;
     double const& electrons_per_oxygenation;
     double const& gbw_canopy;
+    double const& gm_at_25;
+    double const& gm_Ha;
+    double const& gm_Hd;
+    double const& gm_S;
     double const& Gs_min;
     double const& Gstar_at_25;
     double const& Gstar_Ea;
@@ -132,6 +141,7 @@ class c3_leaf_photosynthesis : public direct_module
 
     // Pointers to output quantities
     double* Assim_op;
+    double* Cc_op;
     double* Ci_op;
     double* Cs_op;
     double* EPenman_op;