+++ src/module_library/c4photo.cpp
@@ -126,7 +126,7 @@ photosynthesis_outputs c4photoC(
         // Using Ci and Gs, make a new estimate of the assimilation rate. If
         // the initial value of Ci was correct, this should be identical to
         // Assim.
-        double Gt = sequential_conductance(gbw / dr_boundary, Gs / dr_stomata);  // mol / m^2 / s
+        double Gt = sequential_conductance({gbw / dr_boundary, Gs / dr_stomata});  // mol / m^2 / s
 
         return Gt * (Ca_pa - Ci_pa) / atmospheric_pressure * 1e6 - Assim;  // micromol / m^2 / s
     };
@@ -160,6 +160,7 @@ photosynthesis_outputs c4photoC(
     return photosynthesis_outputs{
         /* .Assim = */ Assim,                       // micromol / m^2 /s
         /* .Assim_conductance = */ an_conductance,  // micromol / m^2 / s
+        /* .Cc = */ 0,                              // micromol / mol
         /* .Ci = */ Ci,                             // micromol / mol
         /* .Cs = */ BB_res.cs,                      // micromol / m^2 / s
         /* .GrossAssim = */ Assim + RT,             // micromol / m^2 / s