+++ src/module_library/conductance_helpers.h
@@ -1,10 +1,11 @@
 #ifndef CONDUCTANCE_HELPERS_H
 #define CONDUCTANCE_HELPERS_H
 
+#include <initializer_list>            // for std::initializer_list
 #include "water_and_air_properties.h"  // for molar_volume
 
 /**
- *  @brief Convert a conductance value from a "molecular" basis (in units of
+ *  @brief Convert a conductance value from a "molar" basis (in units of
  *  mol / m^2 / s) to a "mass" basis (in units of m / s).
  *
  *  There are two different conventions for specifying gas concentrations and
@@ -19,22 +20,23 @@
  *     per area per time (typically kg / m^2 / s). Thus, conductance must be
  *     expressed as length per time (typically m / s).
  *
- *  2. The "molecular" basis: In this convention, gas concentrations are
- *     expressed as a fraction of total molecules (typically mol / mol) and gas
- *     fluxes are expressed as molecules per area per time (typically
- *     mol / m^2 / s). Thus, conductance must be expressed as molecules per
- *     area per time (typically mol / m^2 / s).
+ *  2. The "molar" basis: In this convention, gas concentrations are expressed
+ *     as a fraction of the total number of molecules (typically mol / mol),
+ *     often referred to as "mole fractions", and gas fluxes are expressed as a
+ *     number of molecules per area per time (typically mol / m^2 / s). Thus,
+ *     conductance must be expressed as a number of molecules per area per time
+ *     (typically mol / m^2 / s).
  *
  *  Converting between these conventions is not always straightforward. In the
  *  simple case where temperature and pressure are constant across the gas path,
- *  it can be shown that G_mass = G_molecular * molar_volume, where molar_volume
+ *  it can be shown that G_mass = G_molar * molar_volume, where molar_volume
  *  is the molar volume of an ideal gas at the given temperature and pressure.
  *
  *  We often use this conversion in BioCro, even when temperature is not
  *  constant across the gas path, because errors due to this approximation are
  *  expected to be small.
  *
- *  See `g_to_molecular` for the opposite conversion.
+ *  See `g_to_molar` for the opposite conversion.
  *
  *  @param [in] pressure Pressure along the gas path in Pa.
  *
@@ -56,7 +58,7 @@ inline double g_to_mass(
 
 /**
  *  @brief Convert a conductance value from a "mass" basis (in units of
- *  m / s) to a "molecular" basis (in units of mol / m^2 / s).
+ *  m / s) to a "molar" basis (in units of mol / m^2 / s).
  *
  *  See `g_to_mass` for more information.
  *
@@ -68,7 +70,7 @@ inline double g_to_mass(
  *
  *  @return Conductance along the gas path in mol / m^2 / s.
  */
-inline double g_to_molecular(
+inline double g_to_molar(
     double const pressure,     // Pa
     double const conductance,  // mol / m^2 / s
     double const temperature   // degrees C
@@ -78,42 +80,6 @@ inline double g_to_molecular(
     return conductance / mv_tl;                                // mol / m^2 / s
 }
 
-/**
- *  @brief Calculates the total conductance across two sequential gas paths.
- *
- *  If gas flows from location A to location B and then to location C, there
- *  are two sequential steps: A to B (step 1), and B to C (step 2). Using one
- *  dimensional gas flow equations, we have:
- *
- *  F_1 = G_1 * (C_B - C_A)
- *  F_2 = G_2 * (C_C - C_B)
- *
- *  where F_1/F_2 are fluxes across steps 1 and 2, G_1/G_2 are conductances
- *  across steps 1 and 2, and C_A/C_B/C_C are concentrations at A, B, and C.
- *
- *  At steady state, F_1 = F_2 = F, and the flux across the entire path is
- *  given by
- *
- *  F = G_T * (C_C - C_A)
- *
- *  where G_T is the total conductance across the path. Solving for G_T, we can
- *  find that 1 / G_T = 1 / G_1 + 1 / G_2.
- *
- *  @param [in] conductance_1 Conductance along one step in any conductance
- *              units.
- *
- *  @param [in] conductance_2 Conductance along the other step in the same units
- *              as conductance_1.
- *
- *  @return Total conductance along both steps in the same units as
- *          conductance_1.
- */
-inline double sequential_conductance(
-    double const conductance_1,  // any conductance units
-    double const conductance_2   // same units as conductance_1
-)
-{
-    return 1.0 / (1.0 / conductance_1 + 1.0 / conductance_2);  // same units as conductance_1
-}
+double sequential_conductance(std::initializer_list<double> conductances);
 
 #endif